ファイル分けの整理とextensionを使用してViewControllerをコンパクトにした
