+++ MovieReviewMVP/ReviewMovie/ReviewMoviePresenter.swift
@@ -34,7 +34,7 @@ final class ReviewMoviePresenter : ReviewMoviePresenterInput {
     }
     
     func didTapSaveButton(reviewScore: Double, review: String) {
-        let movieReviewContent = MovieReviewContent(title: movieInfomation.title ?? "", reviewStars: reviewScore, releaseDay: movieInfomation.release_date ?? "", overview: movieInfomation.overview ?? "", review: review, movieImagePath: movieInfomation.poster_path ?? "")
+        let movieReviewContent = MovieReviewElement(title: movieInfomation.title ?? "", reviewStars: reviewScore, releaseDay: movieInfomation.release_date ?? "", overview: movieInfomation.overview ?? "", review: review, movieImagePath: movieInfomation.poster_path ?? "")
         model.saveMovieReview(movieReviewContent)
     }
 }
